doc_rag/pyproject.toml
[project]
name = "doc-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "streamlit>=1.46.1",
    "unstructured>=0.18.1",
    "langchain>=0.3.26",
    "langchain-community>=0.3.26",
    "langchain-openai>=0.3.24",
    "langchain-pinecone>=0.2.8",
    "langchainhub>=0.1.21",
    "langsmith>=0.4.1",
    "python-dotenv>=1.1.0",
    "firecrawl-py>=2.12.0",
]
